split LoginForm into base and index.js/index.native.js files
