+++ src/pages/signin/LoginForm/index.js
@@ -1,0 +1,18 @@
+import React from 'react';
+import BaseLoginForm from './BaseLoginForm';
+import {propTypes, defaultProps} from './loginFormPropTypes';
+
+function LoginForm(props) {
+    return (
+        <BaseLoginForm
+            // eslint-disable-next-line react/jsx-props-no-spreading
+            {...props}
+        />
+    );
+}
+
+LoginForm.displayName = 'LoginForm';
+LoginForm.propTypes = propTypes;
+LoginForm.defaultProps = defaultProps;
+
+export default LoginForm;
